requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Flask>=2.0.0
Pillow>=9.5.0
numpy>=1.24.0
pyttsx3>=2.90
gTTS>=2.3.1
googletrans>=3.1.0a0
playsound>=1.3.0
sympy>=1.12
scipy>=1.10.0
opencv-python>=4.5.0
